\name{mcGlobaloptim-package}
\alias{mcGlobaloptim-package}
\alias{mcGlobaloptim}
\docType{package}
\title{Global optimization using Monte Carlo and Quasi Monte Carlo simultation}
\description{
Global optimization combining Monte Carlo and Quasi Monte Carlo simulation with a local search. 
The local searches can be easily speeded up by using a network of local workstations.}
\details{
\tabular{ll}{
Package: \tab mcGlobaloptim\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2013-10-10\cr
License: \tab GPL-2 | GPL-3\cr
}

}
\author{
Thierry Moudiki
Maintainer: <thierry.moudiki@gmail.com>
}
\references{
C. Dutang, P. Savicky (2013). randtoolbox: Generating and Testing Random Numbers. R package version 1.14. \cr
M. J. A. Eugster, J. Knaus,  C. Porzelius, M. Schmidberger, E. Vicedo (2011).  Hands-on tutorial for parallel computing with R. Computational Statistics. Springer. \cr
M. Gilli, E. Schumann (2010). A Note on 'Good Starting Values' in Numerical Optimisation. Available at SSRN. \cr
F. Glover, G. Kochenberger (2003). Handbook of Metaheuristics. Kluwer Academic Publishers. \cr
H. Niederreiter (1992). Random Number Generation and Quasi-Monte Carlo Methods. Society for Industrial and Applied Mathematics. \cr
M. Schmidberger, M. Morgan, D. Eddelbuettel, H. Yu, L. Tierney, U. Mansmann(2009). State of the art in parallel computing with R. Journal of Statistical Software. \cr
L. Tierney, A. J. Rossini, Na Li and H. Sevcikova (2013). snow: Simple Network of Workstations. R package version 0.3-12. \cr
A. Zhigljavsky, A. Zilinkas (2008). Stochastic Global Optimization. Springer Science+Business Media.
}

